requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.15
redis==5.0.8
requests==2.31.0
tqdm==4.66.5
